\name{Cigar}
\docType{data}
\alias{Cigar}
\title{Cigarette Consumption }
\description{
  a panel of 46 observations  from 1963  to 1992

 \emph{number of observations} :  1380

 \emph{observation} :  regional

   \emph{country} :  United States

}
\usage{data(Cigar)}
\format{A dataframe containing :
  \describe{
    \item{state}{state abbreviation}
    \item{year}{the year}
    \item{price}{price per pack of cigarettes}
    \item{pop}{population}
    \item{pop16}{population above the age of 16}
    \item{cpi}{consumer price index (1983=100)}
    \item{ndi}{per capita disposable income}
    \item{sales}{cigarette sales in packs per capita}
    \item{pimin}{minimum price in adjoining states per pack of cigarettes}
  }
}
\source{
    Baltagi, B.H.  and  D.  Levin (1992) \dQuote{Cigarette taxation: raising revenues and reducing consumption}, \emph{Structural Changes and Economic Dynamics}, \bold{3}, 321--335.

    Baltagi, B.H.,  J.M.   Griffin and  W.  Xiong (2000) \dQuote{To pool or not to pool: homogeneous versus heterogeneous estimators applied to cigarette demand}, \emph{Review of Economics and Statistics}, \bold{82}, 117--126.

}
\references{
    Baltagi, Badi H. (2003) \emph{Econometric analysis of panel data}, John Wiley and sons, \url{http://www.wiley.com/legacy/wileychi/baltagi/}.
}

\keyword{datasets}
