% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_cips.R
\name{cipstest}
\alias{cipstest}
\title{Cross-sectionally Augmented IPS Test for Unit Roots in Panel Models}
\usage{
cipstest(
  x,
  lags = 2,
  type = c("trend", "drift", "none"),
  model = c("cmg", "mg", "dmg"),
  truncated = FALSE,
  ...
)
}
\arguments{
\item{x}{an object of class `"pseries"`,}

\item{lags}{lag order for Dickey-Fuller augmentation,}

\item{type}{one of `"trend"`, `"drift"`, `"none"`,}

\item{model}{one of `"cmg"`, `"mg"`, `"dmg"`,}

\item{truncated}{logical specifying whether to calculate the
truncated version of the test,}

\item{\dots}{further arguments passed to `critvals.cips`
(non-exported function).}
}
\value{
An object of class `"htest"`.
}
\description{
Cross-sectionally augmented Im, Pesaran and Shin (IPS) test for
unit roots in panel models.
}
\details{
This cross-sectionally augmented version of the IPS unit root test
(H0: the `pseries` has a unit root) is a so-called
second-generation panel unit root test: it is in fact robust
against cross-sectional dependence, provided that the default
`type="cmg"` is calculated. Else one can obtain the standard
(`model="mg"`) or cross-sectionally demeaned (`model="dmg"`)
versions of the IPS test.
The critical values and p-values are interpolated from the tables by Pesaran (2007). If the computed statistic is outside the tables, then a warning message is generated.
}
\examples{

data("Produc", package = "plm")
Produc <- pdata.frame(Produc, index=c("state", "year"))
## check whether the gross state product (gsp) is trend-stationary
cipstest(Produc$gsp, type = "trend")

}
\references{
\insertRef{pes07}{plm}
}
\seealso{
[purtest()]
}
\author{
Giovanni Millo
}
\keyword{htest}
