\name{cortab}
\alias{cortab}
\title{Cross--sectional correlation matrix}

\description{
  Computes the cross--sectional correlation matrix
}

\usage{
cortab(x, grouping, groupnames = NULL, value = "statistic", \dots)
}

\arguments{
 \item{x}{an object of class \code{pseries}}
 \item{grouping}{grouping variable,}
 \item{groupnames}{a character vector of group names,}
 \item{value}{to complete}
 \item{\ldots}{further arguments}
}

\value{A matrix}

\keyword{htest}
