\name{Gasoline}
\docType{data}
\alias{Gasoline}
\title{Gasoline Consumption }
\description{
  a panel of 18 observations  from 1960  to 1978

 \emph{number of observations} :  342

 \emph{observation} :  country

   \emph{country} :  OECD

}
\usage{data(Gasoline)}
\format{A dataframe containing :
  \describe{
    \item{country}{a factor with 18 levels}
    \item{year}{the year}
    \item{lgaspcar}{logarithm of motor gasoline consumption per auto}
    \item{lincomep}{logarithm of real per-capita income}
    \item{lrpmg}{logarithm of real motor gasoline price}
    \item{lcarpcap}{logarithm of the stock of cars per capita}
  }
}
\source{
    Baltagi, B.H.  and  Y.J.  Griggin (1983) \dQuote{Gasoline demand in the OECD: an application of pooling and testing procedures}, \emph{European Economic Review}, \bold{22}.

}
\references{
    Baltagi, Badi H. (2003) \emph{Econometric analysis of panel data}, John Wiley and sons, \url{http://www.wiley.com/legacy/wileychi/baltagi/}.
}

\keyword{datasets}
