\name{Grunfeld}
\docType{data}
\alias{Grunfeld}
\title{Grunfeld Investment Data }
\description{
  a panel of 10 observations  from 1935  to 1954

 \emph{number of observations} :  200

 \emph{observation} :  production units

   \emph{country} :  United States

}
\usage{data(Grunfeld)}
\format{A dataframe containing :
  \describe{
    \item{firm}{observation}
    \item{year}{date}
    \item{inv}{gross Investment}
    \item{value}{value of the firm}
    \item{capital}{stock of plant and equipment}
  }
}
\source{
    Moody's Industrial Manual, Survey of Current Business.

}
\references{
    Greene, W.H. (2003) \emph{Econometric Analysis}, Prentice Hall, \url{http://www.prenhall.com/greene/greene1.html}, Table F13.1.

    Baltagi, Badi H. (2003) \emph{Econometric analysis of panel data}, John Wiley and sons, \url{http://www.wiley.com/legacy/wileychi/baltagi/}.
}

\seealso{for the complete Grunfeld's data (11 firms), see
  \code{\link{Grunfeld}}, in the \code{AER} packages.
}
\keyword{datasets}
