\name{Hedonic}
\docType{data}
\alias{Hedonic}
\title{Hedonic Prices of Cencus Tracts in Boston }
\description{
  a cross-section

 \emph{number of observations} :  506

 \emph{observation} :  regional

   \emph{country} :  United States

}
\usage{data(Hedonic)}
\format{A dataframe containing :
  \describe{
    \item{mv}{median value of owner--occupied homes}
    \item{crim}{crime rate}
    \item{zn}{proportion of 25,000 square feet residential lots}
    \item{indus}{proportion of nonretail business acres}
    \item{chas}{is the tract bounds the Charles River ?}
    \item{nox}{annual average nitrogen oxide concentration in parts per hundred million}
    \item{rm}{average number of rooms}
    \item{age}{proportion of owner units built prior to 1940}
    \item{dis}{weighted distances to five employment centers in the Boston area}
    \item{rad}{index of accessibility to radial highways}
    \item{tax}{full value property tax rate (\$/\$10,000)}
    \item{ptratio}{pupil/teacher ratio}
    \item{blacks}{proportion of blacks in the population}
    \item{lstat}{proportion of population that is lower status}
    \item{townid}{town identifier}
  }
}
\source{
    Harrison, D.  and  D.L.  Rubinfeld (1978) \dQuote{Hedonic housing prices and the demand for clean air}, \emph{Journal of Environmental Economics Ans Management}, \bold{5}, 81--102.

    Belsley, D.A.,  E.   Kuh and  R. E.  Welsch (1980) \emph{Regression diagnostics: identifying influencal data ans sources of collinearity}, John Wiley, New--York.

}
\references{
    Baltagi, Badi H. (2003) \emph{Econometric analysis of panel data}, John Wiley and sons, \url{http://www.wiley.com/legacy/wileychi/baltagi/}.
}

\keyword{datasets}
