\name{LaborSupply}
\docType{data}
\alias{LaborSupply}
\title{Wages and Hours Worked }
\description{
  a panel of 532  observations  from 1979  to 1988

 \emph{number of observations} :  5320

}
\usage{data(LaborSupply)}
\format{A dataframe containing :
  \describe{
    \item{lnhr}{log of annual hours worked}
    \item{lnwg}{log of hourly wage}
    \item{kids}{number of children}
    \item{age}{age}
    \item{disab}{bad health}
    \item{id}{id}
    \item{year}{year}
  }
}
\source{
    Ziliak, Jim (1997) \dQuote{Efficient Estimation With Panel Data when Instruments are Predetermined: An Empirical Comparison of Moment-Condition Estimators}, \emph{Journal of Business and Economic Statistics}, \bold{419--431}.

}
\references{
    Cameron, A.C.  and  P.K.  Trivedi (2005) \emph{Microeconometrics : methods and applications}, Cambridge, pp. 708--15, 754--6.

      Journal of Business Economics and Statistics web site : \url{http://www.amstat.org/publications/jbes/}.
}

\keyword{datasets}
