% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plm-package.R
\docType{data}
\name{Parity}
\alias{Parity}
\title{Purchasing Power Parity and other parity relationships}
\format{A data frame containing :
\describe{
\item{country}{country codes: a factor with 17 levels}
\item{time}{the quarter index, 1973Q1-1998Q4}
\item{ls}{log spot exchange rate vs. USD}
\item{lp}{log price level}
\item{is}{short term interest rate}
\item{il}{long term interest rate}
\item{ld}{log price differential vs. USA}
\item{uis}{U.S. short term interest rate}
\item{uil}{U.S. long term interest rate} }}
\source{
\insertRef{COAK:FUER:SMIT:06}{plm}
}
\description{
A panel of 104 quarterly observations from 1973Q1 to 1998Q4
}
\details{
\emph{total number of observations} : 1768

\emph{observation} : country

\emph{country} : OECD
}
\references{
\insertRef{COAK:FUER:SMIT:06}{plm}

\insertRef{DRIS:KRAA:98}{plm}
}
\keyword{datasets}
