% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_cd.R
\name{pcdtest}
\alias{pcdtest}
\alias{pcdtest.formula}
\alias{pcdtest.panelmodel}
\alias{pcdtest.pseries}
\title{Tests of cross-section dependence for panel models}
\usage{
pcdtest(x, ...)

\method{pcdtest}{formula}(x, data, index = NULL, model = NULL,
  test = c("cd", "sclm", "bcsclm", "lm", "rho", "absrho"), w = NULL, ...)

\method{pcdtest}{panelmodel}(x, test = c("cd", "sclm", "bcsclm", "lm", "rho",
  "absrho"), w = NULL, ...)

\method{pcdtest}{pseries}(x, test = c("cd", "sclm", "bcsclm", "lm", "rho",
  "absrho"), w = NULL, ...)
}
\arguments{
\item{x}{an object of class \code{formula}, \code{panelmodel}, or \code{pseries}
(depending on the respective interface) describing the model to
be tested,}

\item{\dots}{further arguments to be passed on to \code{plm}, such as
\code{effect} or \code{random.method}.}

\item{data}{a \code{data.frame},}

\item{index}{an optional numerical index, if \code{NULL}, the first two
columns of the data.frame provided in argument \code{data} are
assumed to be the index variables; for further details see
\code{\link[=pdata.frame]{pdata.frame()}},}

\item{model}{an optional character string indicating which type of
model to estimate; if left to \code{NULL}, the original
heterogeneous specification of Pesaran is used,}

\item{test}{the type of test statistic to be returned. One of
\itemize{ \item \code{"cd"} for Pesaran's CD statistic, \item \code{"lm"}
for Breusch and Pagan's original LM statistic, \item \code{"sclm"}
for the scaled version of Breusch and Pagan's LM statistic,
\item \code{"bcsclm"} for the bias-corrected scaled version of
Breusch and Pagan's LM statistic, \item \code{"rho"} for the average
correlation coefficient, \item \code{"absrho"} for the average
absolute correlation coefficient,}}

\item{w}{either \code{NULL} (default) for the global tests or -- for the
local versions of the statistics -- a \code{n x n} \code{matrix}
describing proximity between individuals, with \eqn{w_ij = a}
where \eqn{a} is any number such that \code{as.logical(a)==TRUE}, if
\eqn{i,j} are neighbours, \eqn{0} or any number \eqn{b} such
that \code{as.logical(b)==FALSE} elsewhere. Only the lower
triangular part (without diagonal) of \code{w} after coercing by
\code{as.logical()} is evaluated for neighbouring information (but
\code{w} can be symmetric). See also \strong{Details} and
\strong{Examples},}
}
\value{
An object of class \code{"htest"}.
}
\description{
Pesaran's CD or Breusch--Pagan's LM (local or global) tests for cross
sectional dependence in panel models
}
\examples{

data("Grunfeld", package = "plm")
## test on heterogeneous model (separate time series regressions)
pcdtest(inv ~ value + capital, data = Grunfeld,
        index = c("firm", "year"))

## test on two-way fixed effects homogeneous model
pcdtest(inv ~ value + capital, data = Grunfeld, model = "within",
        effect = "twoways", index = c("firm", "year"))

## test on panelmodel object
g <- plm(inv ~ value + capital, data = Grunfeld, index = c("firm", "year"))
pcdtest(g)

## scaled LM test
pcdtest(g, test = "sclm")

## test on pseries
pGrunfeld <- pdata.frame(Grunfeld)
pcdtest(pGrunfeld$value)

## local test
## define neighbours for individual 2: 1, 3, 4, 5 in lower triangular matrix
w <- matrix(0, ncol= 10, nrow=10)
w[2,1] <- w[3,2] <- w[4,2] <- w[5,2] <- 1
pcdtest(g, w = w)

}
\references{
\insertRef{BALT:FENG:KAO:12}{plm}

\insertRef{BREU:PAGA:80}{plm}

\insertRef{PESA:04}{plm}

\insertRef{PESA:15}{plm}
}
\keyword{htest}
