% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg-package.R
\name{reexports}
\alias{reexports}
\alias{maxLik}
\alias{loglik}
\title{Functions exported from other packages}
\description{
These functions are imported from other packages and re-exported by
\pkg{plm} to enable smooth use within \pkg{plm}.  Please follow the links to
view the function's original documentation.
}
\details{
\describe{ \itemize{ \item\code{maxLik}: \code{\link[maxLik]{maxLik}} from
package \pkg{maxLik} } }
}
\keyword{internal}
