\name{Snmesp}
\docType{data}
\alias{Snmesp}
\title{Employment and Wage in Spain }
\description{
  a panel of 738 observations  from 1983  to 1990

 \emph{number of observations} :  5904

 \emph{observation} :  firms

   \emph{country} :  Spain

}
\usage{data(Snmesp)}
\format{A dataframe containing :
  \describe{
    \item{firm}{firm index}
    \item{year}{year}
    \item{n}{log of employment}
    \item{w}{log of wages}
    \item{y}{log of real output}
    \item{i}{log of intermediate inputs}
    \item{k}{log of real capital stock}
    \item{f}{real cash flow}
  }
}
\source{

  Alonso--Borrego, C. and  Arellano, M. (1999) Symmetrically
  Normalized Instrumental-Variable Estimation Using Panel Data,
  \emph{Journal of Business and Economic Statistics}, \bold{17(1)},
  36--49.

}
\references{

  Journal of Business Economics and Statistics web site :
  \url{http://www.amstat.org/publications/jbes/}.

}

\keyword{datasets}
