\name{SumHes}
\docType{data}
\alias{SumHes}
\title{The Penn Table }
\description{
  a panel of 125 observations  from 1960  to 1985

 \emph{number of observations} :  3250

 \emph{observation} :  country

   \emph{country} :  World

}
\usage{data(SumHes)}
\format{A dataframe containing :
  \describe{
    \item{year}{the year}
    \item{country}{the country name (factor)}
    \item{opec}{OPEC member ?}
    \item{com}{communist regime ? }
    \item{pop}{country's population (in thousands)}
    \item{gdp}{real GDP per capita (in 1985 US dollars)}
    \item{sr}{saving rate (in percent)}
  }
}
\source{
    Summers, R.  and  A.  Heston (1991) \dQuote{The Penn world table (mark 5): an exanded set of international comparaisons, 1950-1988}, \emph{Quarterly Journal of Economics}, \bold{29}, 229-256.

}
\references{
    Hayashi, F. (2000) \emph{Econometrics}, Princeton University Press, \url{http://www.e.u-tokyo.ac.jp/~hayashi/hayashi_econometrics.htm}, chapter 5, 358-363.

}

\keyword{datasets}
