\name{Wages}
\docType{data}
\alias{Wages}
\title{Panel Datas of Individual Wages }
\description{
  a panel of 595 observations  from 1976  to 1982

 \emph{number of observations} :  4165

 \emph{observation} :  individuals

   \emph{country} :  United States

}
\usage{data(Wages)}
\format{A dataframe containing :
  \describe{
    \item{exp}{years of full-time work experience}
    \item{wks}{weeks worked}
    \item{bluecol}{blue collar ? }
    \item{ind}{works in a manufacturing industry ?}
    \item{south}{resides in the south ?}
    \item{smsa}{resides in a standard metropolitan statistical are ?}
    \item{married}{married ?}
    \item{sex}{a factor with levels (male,female)}
    \item{union}{individual's wage set by a union contract ?}
    \item{ed}{years of education}
    \item{black}{is the individual black ?}
    \item{lwage}{logarithm of wage}
  }
}
\source{
    Cornwell, C.  and  P.  Rupert (1988) \dQuote{Efficient estimation with panel data: an empirical comparison of instrumental variables estimators}, \emph{Journal of Applied Econometrics}, \bold{3}, 149--155.

    Panel study of income dynamics.

}
\references{
    Baltagi, Badi H. (2003) \emph{Econometric analysis of panel data}, John Wiley and sons, \url{http://www.wiley.com/legacy/wileychi/baltagi/}.
  }

\keyword{datasets}
