% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plm-package.R
\docType{data}
\name{Cigar}
\alias{Cigar}
\title{Cigarette Consumption}
\format{A data frame containing :
\describe{
\item{state}{state abbreviation}
\item{year}{the year}
\item{price}{price per pack of cigarettes}
\item{pop}{population}
\item{pop16}{population above the age of 16}
\item{cpi}{consumer price index (1983=100)}
\item{ndi}{per capita disposable income}
\item{sales}{cigarette sales in packs per capita}
\item{pimin}{minimum price in adjoining states per pack of cigarettes}
}}
\source{
Online complements to Baltagi (2001):

\url{http://www.wiley.com/legacy/wileychi/baltagi/}

Online complements to Baltagi (2013):

\url{http://bcs.wiley.com/he-bcs/Books?action=resource&bcsId=4338&itemId=1118672321&resourceId=13452}
}
\description{
a panel of 46 observations from 1963 to 1992
}
\details{
\emph{total number of observations} : 1380

\emph{observation} : regional

\emph{country} : United States
}
\references{
\insertRef{BALT:01}{plm}

\insertRef{BALT:13}{plm}

\insertRef{BALT:LEVI:92}{plm}

\insertRef{BALT:GRIF:XION:00}{plm}
}
\keyword{datasets}
