\name{Cigar}
\docType{data}
\alias{Cigar}
\title{Cigarette Consumption }
\description{
  a panel of 46 observations  from 1963  to 1992

 \emph{total number of observations} :  1380

 \emph{observation} :  regional

 \emph{country} :  United States
}

\usage{data(Cigar)}

\format{A data frame containing :
  \describe{
    \item{state}{state abbreviation}
    \item{year}{the year}
    \item{price}{price per pack of cigarettes}
    \item{pop}{population}
    \item{pop16}{population above the age of 16}
    \item{cpi}{consumer price index (1983=100)}
    \item{ndi}{per capita disposable income}
    \item{sales}{cigarette sales in packs per capita}
    \item{pimin}{minimum price in adjoining states per pack of cigarettes}
  }
}

\source{
Online complements to Baltagi (2001):

 \url{http://www.wiley.com/legacy/wileychi/baltagi/}
 
Online complements to Baltagi (2013):

  \url{http://bcs.wiley.com/he-bcs/Books?action=resource&bcsId=4338&itemId=1118672321&resourceId=13452}
}

\references{
 Baltagi, Badi H. (2001) \emph{Econometric Analysis of Panel Data}, 2nd ed., John Wiley and Sons.
 
 Baltagi, Badi H. (2013) \emph{Econometric Analysis of Panel Data}, 5th ed., John Wiley and Sons.
 
 Baltagi, B.H. and  D. Levin (1992) \dQuote{Cigarette taxation: raising revenues and reducing consumption}, \emph{Structural Changes and Economic Dynamics}, \bold{3}, 321--335.

 Baltagi, B.H., J.M. Griffin and  W. Xiong (2000) \dQuote{To pool or not to pool: homogeneous versus heterogeneous estimators applied to cigarette demand}, \emph{Review of Economics and Statistics}, \bold{82}, 117--126.
}

\keyword{datasets}
