% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_cd.R
\name{cortab}
\alias{cortab}
\title{Cross--sectional correlation matrix}
\usage{
cortab(x, grouping, groupnames = NULL, value = "statistic", ...)
}
\arguments{
\item{x}{an object of class \code{pseries}}

\item{grouping}{grouping variable,}

\item{groupnames}{a character vector of group names,}

\item{value}{to complete}

\item{\dots}{further arguments}
}
\value{
A matrix
}
\description{
Computes the cross--sectional correlation matrix
}
\keyword{htest}
