% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_cd.R
\name{cortab}
\alias{cortab}
\title{Cross--sectional correlation matrix}
\usage{
cortab(x, grouping, groupnames = NULL, value = "statistic", ...)
}
\arguments{
\item{x}{an object of class \code{pseries}}

\item{grouping}{grouping variable,}

\item{groupnames}{a character vector of group names,}

\item{value}{to complete,}

\item{\dots}{further arguments.}
}
\value{
A matrix with average correlation coefficients within a group
(diagonal) and between groups (off-diagonal).
}
\description{
Computes the cross--sectional correlation matrix
}
\examples{

data("Grunfeld", package = "plm")
pGrunfeld <- pdata.frame(Grunfeld)
grp <- c(rep(1, 100), rep(2, 50), rep(3, 50)) # make 3 groups
cortab(pGrunfeld$value, grouping = grp, groupnames = c("A", "B", "C"))

}
\keyword{htest}
