\name{EmplUK}
\docType{data}
\alias{EmplUK}
\title{Employment and Wages in the United Kingdom}

\description{
 An unbalanced panel of 140 observations from 1976 to 1984

 \emph{total number of observations} :  1031

 \emph{observation} :  firms

 \emph{country} :  United Kingdom

}

\usage{data(EmplUK)}

\format{A data frame containing :
  \describe{
    \item{firm}{firm index}
    \item{year}{year}
    \item{sector}{the sector of activity}
    \item{emp}{employment}
    \item{wage}{wages}
    \item{capital}{capital}
    \item{output}{output}
  }
}

\source{
  Arellano, M. and Bond, S. (1991), Some Tests of Specification
  for Panel Data: Monte Carlo Evidence and an Application to Employment Equations, \emph{The Review of Economic Studies}, vol. \bold{58}(2), 1991, pp. 227--297.
}

\keyword{datasets}
