% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plm-package.R
\docType{data}
\name{EmplUK}
\alias{EmplUK}
\title{Employment and Wages in the United Kingdom}
\format{A data frame containing :
\describe{
\item{firm}{firm index}
\item{year}{year}
\item{sector}{the sector of activity}
\item{emp}{employment}
\item{wage}{wages}
\item{capital}{capital}
\item{output}{output}
}}
\source{
\insertRef{AREL:BOND:91}{plm}
}
\description{
An unbalanced panel of 140 observations from 1976 to 1984
}
\details{
\emph{total number of observations} : 1031

\emph{observation} : firms

\emph{country} : United Kingdom
}
\keyword{datasets}
