\name{Gasoline}
\docType{data}
\alias{Gasoline}
\title{Gasoline Consumption}

\description{
 A panel of 18 observations from 1960 to 1978

 \emph{total number of observations} :  342

 \emph{observation} :  country

 \emph{country} :  OECD

}

\usage{data(Gasoline)}

\format{A data frame containing :
  \describe{
    \item{country}{a factor with 18 levels}
    \item{year}{the year}
    \item{lgaspcar}{logarithm of motor gasoline consumption per car}
    \item{lincomep}{logarithm of real per-capita income}
    \item{lrpmg}{logarithm of real motor gasoline price}
    \item{lcarpcap}{logarithm of the stock of cars per capita}
  }
}
\source{
Online complements to Baltagi (2001):

 \url{http://www.wiley.com/legacy/wileychi/baltagi/}
 
Online complements to Baltagi (2013):

  \url{http://bcs.wiley.com/he-bcs/Books?action=resource&bcsId=4338&itemId=1118672321&resourceId=13452}
}

\references{
 Baltagi, Badi H. (2001) \emph{Econometric Analysis of Panel Data}, 2nd ed., John Wiley and Sons.
 
 Baltagi, Badi H. (2013) \emph{Econometric Analysis of Panel Data}, 5th ed., John Wiley and Sons.

 Baltagi, B.H.  and  J.M. Griffin (1983) \dQuote{Gasoline demand in the OECD: An application of pooling and testing procedures}, \emph{European Economic Review}, \bold{22}(2), 117--137.
}

\keyword{datasets}
