% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plm-package.R
\docType{data}
\name{Gasoline}
\alias{Gasoline}
\title{Gasoline Consumption}
\format{
A data frame containing :
\describe{
\item{country}{a factor with 18 levels}
\item{year}{the year}
\item{lgaspcar}{logarithm of motor gasoline consumption per car}
\item{lincomep}{logarithm of real per-capita income}
\item{lrpmg}{logarithm of real motor gasoline price}
\item{lcarpcap}{logarithm of the stock of cars per capita}
}
}
\source{
Online complements to Baltagi (2001):

\url{http://www.wiley.com/legacy/wileychi/baltagi/}

Online complements to Baltagi (2013):

\url{http://bcs.wiley.com/he-bcs/Books?action=resource&bcsId=4338&itemId=1118672321&resourceId=13452}
}
\description{
A panel of 18 observations from 1960 to 1978
}
\details{
\emph{total number of observations} : 342

\emph{observation} : country

\emph{country} : OECD
}
\references{
\insertRef{BALT:01}{plm}

\insertRef{BALT:13}{plm}

\insertRef{BALT:GRIF:83}{plm}
}
\keyword{datasets}
