\name{Grunfeld}
\docType{data}
\alias{Grunfeld}
\title{Grunfeld's Investment Data}

\description{
  A panel of 10 observations from 1935 to 1954

 \emph{total number of observations} :  200

 \emph{observation} :  production units

 \emph{country} :  United States

}

\usage{data(Grunfeld)}

\format{A data frame containing :
  \describe{
    \item{firm}{observation}
    \item{year}{date}
    \item{inv}{gross Investment}
    \item{value}{value of the firm}
    \item{capital}{stock of plant and equipment}
  }
}

\source{
Online complements to Baltagi (2001).

\url{http://www.wiley.com/legacy/wileychi/baltagi/}.
}

\references{
 Baltagi, Badi H. (2001) \emph{Econometric Analysis of Panel Data}, 2nd ed., John Wiley and Sons.
}

\seealso{For the complete Grunfeld data (11 firms), see
  \code{\link{Grunfeld}}, in the \code{AER} package.
}

\keyword{datasets}
