\name{Grunfeld}
\docType{data}
\alias{Grunfeld}
\title{Grunfeld's Investment Data}

\description{
  A panel of 10 observations from 1935 to 1954

 \emph{total number of observations} :  200

 \emph{observation} :  production units

 \emph{country} :  United States

}

\usage{data(Grunfeld)}

\format{A data frame containing :
  \describe{
    \item{firm}{observation}
    \item{year}{date}
    \item{inv}{gross Investment}
    \item{value}{value of the firm}
    \item{capital}{stock of plant and equipment}
  }
}

\source{
Online complements to Baltagi (2001):

 \url{http://www.wiley.com/legacy/wileychi/baltagi/}
 
Online complements to Baltagi (2013):

  \url{http://bcs.wiley.com/he-bcs/Books?action=resource&bcsId=4338&itemId=1118672321&resourceId=13452}
}

\references{
  Baltagi, Badi H. (2001) \emph{Econometric Analysis of Panel Data}, 2nd ed., John Wiley and Sons.
  
  Baltagi, Badi H. (2013) \emph{Econometric Analysis of Panel Data}, 5th ed., John Wiley and Sons.
}

\seealso{For the complete Grunfeld data (11 firms), see
  \code{\link[AER]{Grunfeld}}, in the \code{AER} package.
  
  A website with various versions of the Grunfeld data set can be found at
  \url{http://statmath.wu-wien.ac.at/~zeileis/grunfeld/}.
  
}

\keyword{datasets}
