% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plm-package.R
\docType{data}
\name{Hedonic}
\alias{Hedonic}
\title{Hedonic Prices of Census Tracts in the Boston Area}
\format{A dataframe containing:
\describe{
\item{mv}{median value of owner--occupied homes}
\item{crim}{crime rate}
\item{zn}{proportion of 25,000 square feet residential lots}
\item{indus}{proportion of no--retail business acres}
\item{chas}{is the tract bounds the Charles River?}
\item{nox}{annual average nitrogen oxide concentration in parts per hundred million}
\item{rm}{average number of rooms}
\item{age}{proportion of owner units built prior to 1940}
\item{dis}{weighted distances to five employment centers in the Boston area}
\item{rad}{index of accessibility to radial highways}
\item{tax}{full value property tax rate ($/$10,000)}
\item{ptratio}{pupil/teacher ratio}
\item{blacks}{proportion of blacks in the population}
\item{lstat}{proportion of population that is lower status}
\item{townid}{town identifier} }}
\source{
Online complements to Baltagi (2001):

\url{http://www.wiley.com/legacy/wileychi/baltagi/}

Online complements to Baltagi (2013):

\url{http://bcs.wiley.com/he-bcs/Books?action=resource&bcsId=4338&itemId=1118672321&resourceId=13452}
}
\description{
A cross-section
}
\details{
\emph{number of observations} : 506

\emph{observation} : regional

\emph{country} : United States
}
\references{
\insertRef{BALT:01}{plm}

\insertRef{BALT:13}{plm}

\insertRef{BESL:KUH:WELS:80}{plm}

\insertRef{HARR:RUBI:78}{plm}
}
\keyword{datasets}
