% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plm-package.R
\docType{data}
\name{LaborSupply}
\alias{LaborSupply}
\title{Wages and Hours Worked}
\format{A data frame containing :
\describe{
\item{lnhr}{log of annual hours worked}
\item{lnwg}{log of hourly wage}
\item{kids}{number of children}
\item{age}{age}
\item{disab}{bad health}
\item{id}{id}
\item{year}{year}
}}
\source{
Online complements to Ziliak (1997).

Journal of Business Economics and Statistics web site:
\url{http://amstat.tandfonline.com/loi/ubes20/}.
}
\description{
A panel of 532 observations from 1979 to 1988
}
\details{
\emph{number of observations} : 5320
}
\references{
\insertRef{CAME:TRIV:05}{plm}

\insertRef{ZILI:97}{plm}
}
\keyword{datasets}
