% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plm-package.R
\docType{data}
\name{Males}
\alias{Males}
\title{Wages and Education of Young Males}
\format{A data frame containing :
\describe{
\item{nr}{identifier}
\item{year}{year}
\item{school}{years of schooling}
\item{exper}{years of experience (computed as \code{age-6-school})}
\item{union}{wage set by collective bargaining?}
\item{ethn}{a factor with levels \verb{black, hisp, other}}
\item{married}{married?}
\item{health}{health problem?}
\item{wage}{log of hourly wage}
\item{industry}{a factor with 12 levels}
\item{occupation}{a factor with 9 levels}
\item{residence}{a factor with levels \verb{rural_area, north_east, northern_central, south}}
}}
\source{
Journal of Applied Econometrics data archive
\url{http://qed.econ.queensu.ca/jae/1998-v13.2/vella-verbeek/}.
}
\description{
A panel of 545 observations from 1980 to 1987
}
\details{
\emph{total number of observations} : 4360

\emph{observation} : individuals

\emph{country} : United States
}
\references{
\insertRef{VELL:VERB:98}{plm}

\insertRef{VERB:04}{plm}
}
\keyword{datasets}
