% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R, R/detect_lin_dep_alias.R
\name{plm-deprecated}
\alias{plm-deprecated}
\alias{pvcovHC}
\alias{plm.data}
\alias{dynformula}
\alias{formula.dynformula}
\alias{print.dynformula}
\alias{pFormula}
\alias{as.Formula.pFormula}
\alias{model.frame.pFormula}
\alias{model.matrix.pFormula}
\alias{detect_lin_dep}
\title{Deprecated functions of plm}
\usage{
pvcovHC(x, ...)

plm.data(x, indexes = NULL)

dynformula(formula, lag.form = NULL, diff.form = NULL, log.form = NULL)

\method{formula}{dynformula}(x, ...)

\method{print}{dynformula}(x, ...)

pFormula(object)

\method{as.Formula}{pFormula}(x, ...)

\method{as.Formula}{pFormula}(x, ...)

\method{model.frame}{pFormula}(formula, data, ..., lhs = NULL, rhs = NULL)

\method{model.matrix}{pFormula}(
  object,
  data,
  model = c("pooling", "within", "Between", "Sum", "between", "mean", "random", "fd"),
  effect = c("individual", "time", "twoways", "nested"),
  rhs = 1,
  theta = NULL,
  cstcovar.rm = NULL,
  ...
)

detect_lin_dep(object, ...)
}
\arguments{
\item{\dots}{further arguments.}

\item{indexes}{a vector (of length one or two) indicating the (individual
and time) indexes (see Details);}

\item{formula}{a formula,}

\item{lag.form}{a list containing the lag structure of each variable in the
formula,}

\item{diff.form}{a vector (or a list) of logical values indicating whether
variables should be differenced,}

\item{log.form}{a vector (or a list) of logical values indicating whether
variables should be in logarithms.}

\item{object, x}{an object of class \code{"plm"},}

\item{data}{a \code{data.frame},}

\item{lhs}{see Formula}

\item{rhs}{see Formula}

\item{model}{see plm}

\item{effect}{see plm}

\item{theta}{the parameter of transformation for the random effect model}

\item{cstcovar.rm}{remove the constant columns or not}
}
\description{
\code{dynformula}, \code{pht}, \code{plm.data}, and \code{pvcovHC} are
deprecated functions which could be removed from \pkg{plm} in a near future.
}
\details{
\code{dynformula} was used to construct a dynamic formula which was the
first argument of \code{pgmm}. \code{pgmm} uses now multi-part formulas.

\code{pht} estimates the Hausman-Taylor model, which can now be estimated
using the more general \code{plm} function.

\code{plm.data} is replaced by \code{pdata.frame}.

\code{pvcovHC} is replaced by \code{vcovHC}.

\code{detect_lin_dep} is replaced by \code{detect.lindep}.
}
