\name{Produc}
\docType{data}
\alias{Produc}

\title{US States Production}

\description{
A panel of 48 observations from 1970 to 1986

 \emph{total number of observations} :  816

 \emph{observation} :  regional

 \emph{country} :  United States
}

\usage{data(Produc)}

\format{A data frame containing :
  \describe{
    \item{state}{the state}
    \item{year}{the year}
    \item{pcap}{public capital stock}
    \item{hwy}{highway and streets}
    \item{water}{water and sewer facilities}
    \item{util}{other public buildings and structures}
    \item{pc}{private capital stock}
    \item{gsp}{gross state products}
    \item{emp}{labor input measured by the employement in non--agricultural payrolls}
    \item{unemp}{state unemployment rate}
  }
}

\source{
Online complements to Baltagi (2001):

 \url{http://www.wiley.com/legacy/wileychi/baltagi/}
 
Online complements to Baltagi (2013):

  \url{http://bcs.wiley.com/he-bcs/Books?action=resource&bcsId=4338&itemId=1118672321&resourceId=13452}
}

\references{
Baltagi, Badi H. (2001) \emph{Econometric Analysis of Panel Data}, 2nd ed., John Wiley and Sons.
  
Baltagi, Badi H. (2013) \emph{Econometric Analysis of Panel Data}, 5th ed., John Wiley and Sons.

Baltagi, B. H.  and  N.  Pinnoi (1995) \dQuote{Public capital stock and state productivity growth: further evidence}, \emph{Empirical Economics}, \bold{20}, 351--359.

Munnell, A.H. (1990) \dQuote{Why has productivity growth declined? Productivity and public investment}, \emph{New England Economic Review}, 3--22.
}

\keyword{datasets}
