% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plm-package.R
\docType{data}
\name{Produc}
\alias{Produc}
\title{US States Production}
\format{A data frame containing :
\describe{
\item{state}{the state}
\item{year}{the year}
\item{region}{the region}
\item{pcap}{public capital stock}
\item{hwy}{highway and streets}
\item{water}{water and sewer facilities}
\item{util}{other public buildings and structures}
\item{pc}{private capital stock}
\item{gsp}{gross state product}
\item{emp}{labor input measured by the employment in non--agricultural payrolls}
\item{unemp}{state unemployment rate} }}
\source{
Online complements to Baltagi (2001):

\url{http://www.wiley.com/legacy/wileychi/baltagi/}

Online complements to Baltagi (2013):

\url{http://bcs.wiley.com/he-bcs/Books?action=resource&bcsId=4338&itemId=1118672321&resourceId=13452}
}
\description{
A panel of 48 observations from 1970 to 1986
}
\details{
\emph{total number of observations} : 816

\emph{observation} : regional

\emph{country} : United States
}
\references{
\insertRef{BALT:01}{plm}

\insertRef{BALT:13}{plm}

\insertRef{BALT:PINN:95}{plm}

\insertRef{MUNN:90}{plm}
}
\keyword{datasets}
