% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plm-package.R
\name{re-export_functions}
\alias{re-export_functions}
\alias{maxLik}
\title{Functions exported from other packages}
\description{
These functions are imported from other packages and re-exported by
\pkg{plm} to enable smooth use within \pkg{plm}.  Please follow the
links to view the function's original documentation.
}
\keyword{internal}
