% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plm-package.R
\docType{data}
\name{Snmesp}
\alias{Snmesp}
\title{Employment and Wages in Spain}
\format{
A data frame containing:

\describe{
\item{firm}{firm index}
\item{year}{year}
\item{n}{log of employment}
\item{w}{log of wages}
\item{y}{log of real output}
\item{i}{log of intermediate inputs}
\item{k}{log of real capital stock}
\item{f}{real cash flow} }
}
\source{
Journal of Business Economics and Statistics data archive:

\url{https://amstat.tandfonline.com/loi/ubes20/}.
}
\description{
A panel of 738 observations from 1983 to 1990
}
\details{
\emph{total number of observations}: 5904

\emph{observation}: firms

\emph{country}: Spain
}
\references{
\insertRef{ALON:AREL:99}{plm}
}
\keyword{datasets}
