% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plm-package.R
\docType{data}
\name{SumHes}
\alias{SumHes}
\title{The Penn World Table, v. 5}
\format{
A data frame containing :
\describe{
\item{year}{the year}
\item{country}{the country name (factor)}
\item{opec}{OPEC member?}
\item{com}{communist regime? }
\item{pop}{country's population (in thousands)}
\item{gdp}{real GDP per capita (in 1985 US dollars)}
\item{sr}{saving rate (in percent)}}
}
\source{
Online supplements to Hayashi (2000).

\url{http://fhayashi.fc2web.com/datasets.htm}
}
\description{
A panel of 125 observations from 1960 to 1985
}
\details{
\emph{total number of observations} : 3250

\emph{observation} : country

\emph{country} : World
}
\references{
\insertRef{HAYA:00}{plm}

\insertRef{SUMM:HEST:91}{plm}
}
\keyword{datasets}
