% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tool_vcovG.R
\name{vcovNW}
\alias{vcovNW}
\alias{vcovNW.plm}
\title{\insertCite{NEWE:WEST:87;textual}{plm} Robust Covariance Matrix
Estimator}
\usage{
vcovNW(x, ...)

\method{vcovNW}{plm}(x, type = c("HC0", "sss", "HC1", "HC2", "HC3",
  "HC4"), maxlag = NULL, wj = function(j, maxlag) 1 - j/(maxlag + 1),
  ...)
}
\arguments{
\item{x}{an object of class \code{"plm"} or \code{"pcce"}}

\item{\dots}{further arguments}

\item{type}{the weighting scheme used, one of \code{"HC0"}, \code{"sss"},
\code{"HC1"}, \code{"HC2"}, \code{"HC3"}, \code{"HC4"}, see Details,}

\item{maxlag}{either \code{NULL} or a positive integer specifying the
maximum lag order before truncation}

\item{wj}{weighting function to be applied to lagged terms,}
}
\value{
An object of class \code{"matrix"} containing the estimate of
the covariance matrix of coefficients.
}
\description{
Nonparametric robust covariance matrix estimators \emph{a la Newey
and West} for panel models with serial correlation.
}
\details{
\code{vcovNW} is a function for estimating a robust covariance matrix of
parameters for a panel model according to the
\insertCite{NEWE:WEST:87;textual}{plm} method.  The function works
as a restriction of the \insertCite{DRIS:KRAA:98;textual}{plm} covariance (see
\code{\link[=vcovSCC]{vcovSCC()}}) to no cross--sectional correlation.

Weighting schemes specified by \code{type} are analogous to those in
\code{\link[sandwich:vcovHC]{sandwich::vcovHC()}} in package \CRANpkg{sandwich} and are
justified theoretically (although in the context of the standard
linear model) by \insertCite{MACK:WHIT:85;textual}{plm} and
\insertCite{CRIB:04;textual}{plm} \insertCite{@see @ZEIL:04}{plm}.

The main use of \code{vcovNW} is to be an argument to other functions,
e.g.  for Wald--type testing: argument \code{vcov.} to \code{coeftest()},
argument \code{vcov} to \code{waldtest()} and other methods in the
\CRANpkg{lmtest} package; and argument \code{vcov.} to
\code{linearHypothesis()} in the \CRANpkg{car} package (see the
examples). Notice that the \code{vcov} and \code{vcov.} arguments allow to
supply a function (which is the safest) or a matrix
\insertCite{@see @ZEIL:04, 4.1-2 and examples below}{plm}.
}
\examples{

library(lmtest)
library(car)
data("Produc", package="plm")
zz <- plm(log(gsp)~log(pcap)+log(pc)+log(emp)+unemp, data=Produc, model="pooling")
## standard coefficient significance test
coeftest(zz)
## NW robust significance test, default
coeftest(zz, vcov.=vcovNW)
## idem with parameters, pass vcov as a function argument
coeftest(zz, vcov.=function(x) vcovNW(x, type="HC1", maxlag=4))
## joint restriction test
waldtest(zz, update(zz, .~.-log(emp)-unemp), vcov=vcovNW)
## test of hyp.: 2*log(pc)=log(emp)
linearHypothesis(zz, "2*log(pc)=log(emp)", vcov.=vcovNW)

}
\references{
\insertRef{CRIB:04}{plm}

\insertRef{DRIS:KRAA:98}{plm}

\insertRef{MACK:WHIT:85}{plm}

\insertRef{NEWE:WEST:87}{plm}

\insertRef{ZEIL:04}{plm}
}
\seealso{
\code{\link[sandwich:vcovHC]{sandwich::vcovHC()}} from the \CRANpkg{sandwich} package
for weighting schemes (\code{type} argument).
}
\author{
Giovanni Millo
}
\keyword{regression}
